Refactored timer logging to TimeRecorder and added timer stats publishing.
All timer logging has been moved to the TimeRecorder interface and
implementations written for the existing recorders.  The new interface
allows more storage locations (database, redis, etc) and the ability to
write an
off-line
recorder.

The TimeStatsPublisher is used to provide a store for the new
TimerStatsPublisherAction that will publish new TimerBasicStatistics as
they are updated.